[IMP] mail: add bounce fake reference in catchall bounce emails
Purpose: be able to break loops due to replies to bounce emails which
in turn trigger a bounce.

Specification: use the detection loop tag already used in loop
detection. It consists in adding a fake reference to bounce email
references, that is found back if someone replies to it. We then
detect someone replied to a bounce email and avoid replying to it
again.

Task-3895869

Part-of: odoo/odoo#187696
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>