[FIX] account: make accounting fields not required
When having Accounting installed with a company without chart template,
we can not create a partner as bot `property_account_payable_id` and
`property_account_receivable_id` are required.

With this commit we add a non stored field with default value to handle the
`required` attribute on the partner view.

opw-4323694

closes odoo/odoo#189341

Related: odoo/enterprise#74994
Signed-off-by: John Laterre (jol) <jol@odoo.com>